Project/noise2noise/requirements_video.txt
absl-py==0.1.9
appdirs==1.4.3
appnope==0.1.0
bleach==3.3.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator
dj-static==0.0.6
django-analytical==2.4.0
django-appconf==1.0.2
django-db-logger==0.1.6
djangorestframework==3.11.2
docopt==0.6.2
ElementTreeFactory==1.0
entrypoints==0.2.3
enum34==1.1.6
first==2.0.1
futures==3.1.1
gulpio==521.36
gunicorn==19.7.1
html5lib==0.9999999
idna==2.6
imageio
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
jedi==0.11.1
Jinja2==2.11.3
joblib==0.11
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.3
Markdown==2.6.11
MarkupSafe==1.1.0
matplotlib==2.1.2
mistune==0.8.3
moviepy==0.2.3.2
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.4.5
notebook==6.1.5
numpy==1.14.0
oauthlib==2.0.6
olefile==0.45.1
opencv-python==3.4.0.14
packaging==16.8
pafy==0.5.3.1
pandas==0.23.0
pandocfilters==1.4.2
parso==0.1.1
patsy==0.5.0
pexpect==4.3.1
pickleshare==0.7.4
Pillow==6.2.0
pip-tools==1.11.0
prompt-toolkit==1.0.15
protobuf==3.5.1
psycopg2==2.7.3.2
ptyprocess==0.5.2
pycparser==2.18
pycurl==7.43.0.1
pydub==0.20.0
pygame==1.9.4
Pygments==2.7.4
pymongo==3.6.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==5.4
pyzmq
qtconsole==4.3.1
requests==2.20.0
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.4.2
sh==1.12.14
simplegeneric==0.8.1
sklearn==0.0
sklearn-extras==0.1.1
sklearn-pandas==1.6.0
static3==0.7.0
statsmodels==0.8.0
tensorflow==1.15.0
tensorflow-tensorboard==1.5.0
terminado==0.8.1
testpath==0.3.1
tornado
tqdm
traitlets==4.3.2
urllib2-file==0.2.1
urllib2-prior-auth==0.2.0
urllib3==1.24.2
video2tfrecord==0.2
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==3.3.1
widgetsnbextension==3.1.3